lua: Fix [bug] not quitting to the main screen with defeat on 1.14
Because of *some* behaviour change made since 1.12, [endlevel] not only
doesn't immediately have an effect in this context (probably because
global [lua] runs *before* the story screen instead of after), but it
also doesn't signal defeat when used in what turns out to be the
"_from_lua" event context, instead signaling victory and the completion
of the campaign (bizarrely enough).

Deferring the [endlevel] action to run on the preload event if we find
ourselves in this situation (e.g. for the version checks) works,
although the player still gets to see the story screen first.