lua: Add version compatibility check module
This is a very simplified rewrite of the code previously implemented
independently by IftU and AtS. Notably, it only considers support for a
single branch, and it doesn't have a way to outright prevent running the
campaign on incompatible versions other than the ones provided in a
blacklist.

It does introduce a behaviour change that's been requested by vultraz a
lot though, which is that in maintainer mode the experimental version
support notice is skipped.