lua/common: Reconcile AtS's version of [setup_doors] with IftU's
Along with a couple of minor improvements.

The semantics of the IftU version are required by IftU S22A, as
otherwise it is impossible to set up the doors required for a crucial
gameplay event (as found by nemaara). The AtS provision of requiring
target locations to be vacant is a bit confusing and I don't remember if
there's a specific reason for it (not clobbering Shaxthal Worms?) but it
turns out to be trivial to reconcile it with the ability to provide a
custom SLF.