Optimize global sound sources' performance by making them actually global
While binding them to a fixed location and setting their volume ranges
to ludicrously high values is technically okay, it turns out it results
in the game pointlessly calling sound::reposition_sound() every time the
viewport position changes, when normally
soundsource::positional_source::update_positions() would skip that step
and the distance calculation entirely if the source is *truly* global,
i.e. not bound to any location sets.

As it turns out, the sound source feature is still largely undocumented
and this little fact isn't even in the wiki yet. Ah, the wonders of
being a mainline developer.