Use UNIT_SPEAKS_FOR_UNDEAD_MINION for item messages
Because this macro is shared between IftU and all episodes of AtS, we
can't make the same assumption as we did in IftU and fallback to Mal
Keshar -- he's not present at all after AtS E1. Instead we fall back to
the controlling side's first leader instead (Galas or Elynia), which
makes just as much sense IMO.

We also don't need to scroll back to the unit or its leader every time,
since this sequence is run in a moveto event handler so we can expect
the viewport to be at the right position already (or close enough). This
avoids annoying scrolls when we fall back to a leader that isn't
anywhere near the primary unit.