lua/gui/amla_list: Fix and improve amla:unit_can_get()
Fixed the require_amla logic not working, and made it so the reason for
a false value is returned as well so the UI can format entries
differently depending on whether they are being disabled by exclude_amla
or require_amla. This means prioritizing exclude_amla in the logic over
require_amla, since it has greater significance to the player.