lua/gui/amla_list: Finish up implementation
Now a WML menu item that executes this command is injected into the
gamestate unconditionally during initialization. The display conditions
are written specifically to avoid accidentally spoilering certain unit
types that may be controlled by enemy AIs, and also to avoid having to
write specific filters for hero units in IftU and AtS. This makes it a
fast and tidy drop-in solution basically.