Implement randomly generated names for Verlissh components
This is a bit of an overengineered item due to the context-free grammar
implementation mysteriously not randomizing non-terminals enough,
resulting in visibly repetitive sequences of numbers and letters in each
output. Instead of generating those with the name generator, we have the
name generator give us a template that's then filled in in a later step
using Lua.

In all cases, Lua has to be used because it turns out Wesnoth doesn't
have support for setting the name generator settings on a per-unit-type
basis. We don't want an extra race for the Verlissh just to give matrix
parts their own names, especially since that's not even going to be used
in IftU anyway; so we piggyback on the unit placed event instead.