Rework [color_adjust]-based screen effects to use [screen_fade]
This includes making it so [screen_fade] and the new [cloak_screen]
action are both skippable by pressing Escape.

Additionally, this adds a workaround for an issue where using
[screen_fade] to pre-set the screen fade effect before the start event
begins like we used to do with [color_adjust] is impossible. We work
this around by using [color_adjust] like before in this particular
case and hooking into the start event as soon as possible to swap the
[color_adjust] with a [screen_fade] effect. So far this seems to work
even in the more involved cases of screen fade effects in AtS (E2S0,
E3S0, E3S1, E3S8C tested).