lua/cloak_screen: Fix [color_adjust] parameters to make sense
Pure black and pure white cannot be implemented by using those literal
values in [color_adjust], since it is meant to be a relative ToD
lighting shift rather than a solid colour effect unlike [screen_fade].

This change makes it so we only use the [color_adjust] workaround when
the input is pure black or pure white, and in those cases we use
either -512,-512,-512 or +512,+512,+512 shifts.

Fixes scenarios such as AtS E2S0 and E3S9 displaying a double fade-in
at the start.