lua/journeylog: New JourneyLog system
This implements a mechanism to record every [message] that ends up
being displayed to the player, into a database than can be browsed via
a user interface accessed through the game's context menu.

The goal is to extend this mechanism to allow adding "lore" entries as
well that can be acquired by the player for later perusal, e.g. by
triggering optional events in scenarios. The general public-facing API
is still a work-in-progress, but the goal is to eventually be able to
ship this as a standalone game modification that can be used with any
campaign, not just in Naia campaigns.

For now, however, Naia and JourneyLog have a symbiotic relationship in
the form of them being aware of each other patching certain critical
game functionality and needing to work around that. Hopefully I can
figure out an easy way to allow JourneyLog to be optionally extracted
out of Naia without having to provide a whole repository for it alone.