lua/journeylog: Unify, simplify, and improve in-line portrait display
This makes it so the same canvas definition is used both for bio
portrait displays and dialog portrait displays, which are now also
displayed in the form of interactive buttons.

Additionally, now we use different scaling strategies depending on
whether an image is pixel art or not, by assuming that anything whose
path starts with 'portraits/' is not pixel art. Previously we assumed
pixel art everywhere for dialog portraits and non-pixel art for bios.