lua/gui/campaign_intro: Another major rework
This version of the campaign intro UI merges the functionality from the
1st character help prompt in IftU scenario 1 into the intro UI proper,
in the form of an optional control. The intro UI now also supports more
than one page, with the player being required to click on a button to
advance to the next page.

(Currently it's possible to use the keyboard to skip the full intro UI,
which may result in option variables being left in an undefined state.)