Rework how Galas's AMLA tree is implemented
This makes it so AMLAs are included for every single member of the unit
tree for two reasons:

 1. [amla_list] needs to be able to see all possible AMLAs at any given
    time, and the best way to achieve this is to include the AMLAs in
    the unit type definitions rather than make an exception in its
    implementation wherein it inspects unit types other than the unit's
    active one.

 2. This ensures that the logical grouping of the AMLAs in the Advance
    Unit dialog stays consistent across all levels, rather than having
    the order reflect the position in the WML with the recursive
    inclusion system.

This commit silently fixes a couple of cases where AMLAs had the wrong
unit promotion requirements because of me changing my mind about this
same implementation shortly before committing the version that I'm
replacing again.

Additionally, this version properly labels the Promotion advancements
so they are more easily recognisable, e.g. in [amla_list].