lua/patch: Fix major mishap with WML variables in [message]
A complete misuse of wml.literal here was causing WML variable usage in
[message] [option] in particular to break completely, since wml.literal
intentionally evaluates everything immediately... including variables
which don't exist yet, or whose value will change during the processing
of [option].

The implications for this are quite grave and only mitigated by the
fact that usage of [message] [option] in Ethea phase 1 is VERY sparse
and mostly uncomplicated. However, this caused extremely hard-to-debug
issues while testing new code in IftU scenario 4.